added SessionService interface to create absrtaction between impl and calling code to aid separating into separate service
